+++ lib/util/pdf_object_helper.py
@@ -10,6 +10,7 @@ from lib.util.logging import log
 
 # In the case of easy key/value pairs the reference_key and the reference_address are the same but
 # for more complicated references the reference_address will be the reference_key plus sub references.
+#
 #   e.g. a reference_key for a /Font labeled /F1 could be '/Resources' but the reference_address
 # might be '/Resources[/Font][/F1] if the /Font is a directly embedded reference instead of a remote one.
 PdfObjectRef = namedtuple('PdfObjectRef', ['reference_key', 'reference_address', 'pdf_obj'])
@@ -21,11 +22,6 @@ def pdf_object_id(pdf_object):
         return pdf_object.idnum
     else:
         return None
-
-
-def does_list_have_any_references(_list):
-    """Return true if any element of _list is an IndirectObject"""
-    return any(isinstance(item, IndirectObject) for item in _list)
 
 
 def get_references(obj: PdfObject, ref_key=None, ref_address=None) -> [PdfObjectRef]:
@@ -45,15 +41,18 @@ def get_references(obj: PdfObject, ref_key=None, ref_address=None) -> [PdfObject
             if not isinstance(element, (IndirectObject, list, dict)):
                 continue
 
-            _ref_key = ref_key or i
             _ref_address = f"[{i}]" if ref_address is None else f"{ref_address}[{i}]"
-            return_list += get_references(element, _ref_key, _ref_address)
+            return_list += get_references(element, ref_key or i, _ref_address)
     elif isinstance(obj, dict):
         for k, v in obj.items():
-            _ref_key = ref_key or k
             _ref_address = k if ref_address is None else f"{ref_address}[{k}]"
-            return_list += get_references(v, _ref_key, _ref_address)
+            return_list += get_references(v, ref_key or k, _ref_address)
     else:
-        log.debug(f"Adding no references for {obj}")
+        log.debug(f"Adding no references for PdfObject reference '{ref_key}' -> '{obj}'")
 
     return return_list
+
+
+def does_list_have_any_references(_list):
+    """Return true if any element of _list is an IndirectObject"""
+    return any(isinstance(item, IndirectObject) for item in _list)